requirements/requirements.txt
numpy
hydra-core
tensorboard
torch
torchvision
omegaconf
scipy
pulp
sklearn
matplotlib
wandb
tqdm
